Remove TypeScript, reorganize directory layout
- Remove all TypeScript and JavaScript code
- Remove all TypeScript- and JavaScript-related config files
- Remove all TypeScript- and JavaScript-related rules from
  .gitignore
- Remove npm- and emulator-related configuration from Firebase
  config files
- Rename the Go "app" module to "functions"
- Remove ability to read TCN reports from Firestore
