--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-  "javascript.implicitProjectConfig.experimentalDecorators": true
-}
