+++ functions/cmd/main.go
@@ -4,13 +4,13 @@ import (
 	"log"
 	"os"
 
-	"app"
+	"functions"
 
 	"github.com/GoogleCloudPlatform/functions-framework-go/funcframework"
 )
 
 func main() {
-	funcframework.RegisterHTTPFunction("/challenge", app.ChallengeHandler)
+	funcframework.RegisterHTTPFunction("/challenge", functions.ChallengeHandler)
 	// Use PORT environment variable, or default to 8080.
 	port := "8088"
 	if envPort := os.Getenv("PORT"); envPort != "" {
