+++ functions/internal/util/util.go
