--- functions/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "functions",
-  "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "tsc": "tsc --build tsconfig.json",
-    "ts-babel": "npx babel src --out-dir lib --extensions '.ts,.tsx'",
-    "build": "npm run tsc && npm run ts-babel",
-    "serve": "npm run build && DEBUG_LOGGING=1 firebase emulators:start",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log",
-    "pretest": "tsc --build tsconfig.test.json",
-    "test": "npm run build && env FIRESTORE_EMULATOR_HOST=localhost:8080 mocha --timeout 5000 --ui mocha-typescript .mocha/test"
-  },
-  "engines": {
-    "node": "10"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "firebase-admin": "^8.6.0",
-    "firebase-functions": "^3.3.0",
-    "sparkson": "^1.3.3",
-    "tcn-node": "github:covid19risk/tcn-node"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.8.4",
-    "@babel/core": "^7.9.0",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/plugin-proposal-decorators": "^7.8.3",
-    "@babel/preset-env": "^7.9.5",
-    "@babel/preset-typescript": "^7.9.0",
-    "@firebase/testing": "^0.19.0",
-    "@types/chai": "^4.2.11",
-    "@types/mocha": "^7.0.2",
-    "babel-plugin-transform-typescript-metadata": "^0.3.0",
-    "chai": "^4.2.0",
-    "chai-http": "^4.3.0",
-    "firebase-functions-test": "^0.1.6",
-    "mocha": "^7.1.1",
-    "mocha-typescript": "^1.1.17",
-    "tslint": "^5.20.1",
-    "typescript": "^3.8.3"
-  },
-  "private": true
-}
