--- functions/src/tcn.d.ts
@@ -1,1 +1,0 @@
-declare module 'tcn-node';
