--- functions/src/types/serverResponse.ts
@@ -1,5 +1,0 @@
-export interface Root {
-    status: string;
-    message: string;
-    data: string[];
-}
