--- functions/src/types/signedReportRequest.ts
@@ -1,9 +1,0 @@
-export interface Root {
-    temporary_contact_key_bytes:          string;
-    memo_data:                            string;
-    memo_type:                            number;
-    start_index:                          number;
-    end_index:                            number;
-    signature_bytes:                      string;
-    report_verification_public_key_bytes: string;
-}
