--- functions/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.test.json",
-  "compilerOptions": {
-    "outDir": "lib",
-    "experimentalDecorators": true,
-    "noEmit": true,
-    "noImplicitAny": true
-  },
-  "exclude": ["test"],
-  "include": ["src/**/*.ts"]
-}
