--- functions/tsconfig.test.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "sourceMap": true,
-    "strict": true,
-    "target": "es2017",
-    "experimentalDecorators": true,
-    "outDir": ".mocha"
-  },
-  "compileOnSave": true,
-  "include": ["test/**/*.ts"]
-}
